+++ java/src/main/java/ai/rapids/cudf/ParquetOptions.java
@@ -18,6 +18,10 @@
 
 package ai.rapids.cudf;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
 /**
  * Options for reading a parquet file
  */
@@ -26,14 +30,23 @@ public class ParquetOptions extends ColumnFilterOptions {
   public static ParquetOptions DEFAULT = new ParquetOptions(new Builder());
 
   private final DType unit;
+  private final boolean[] readBinaryAsString;
 
   private ParquetOptions(Builder builder) {
     super(builder);
     unit = builder.unit;
+    readBinaryAsString = new boolean[builder.binaryAsStringColumns.size()];
+    for (int i = 0 ; i < builder.binaryAsStringColumns.size() ; i++) {
+      readBinaryAsString[i] = builder.binaryAsStringColumns.get(i);
+    }
   }
 
   DType timeUnit() {
     return unit;
+  }
+
+  boolean[] getReadBinaryAsString() {
+    return readBinaryAsString;
   }
 
   public static Builder builder() {
@@ -42,6 +55,7 @@ public class ParquetOptions extends ColumnFilterOptions {
 
   public static class Builder extends ColumnFilterOptions.Builder<Builder> {
     private DType unit = DType.EMPTY;
+    final List<Boolean> binaryAsStringColumns = new ArrayList<>();
 
     /**
      * Specify the time unit to use when returning timestamps.
@@ -51,6 +65,41 @@ public class ParquetOptions extends ColumnFilterOptions {
     public Builder withTimeUnit(DType unit) {
       assert unit.isTimestampType();
       this.unit = unit;
+      return this;
+    }
+
+    /**
+     * Include one or more specific columns.  Any column not included will not be read.
+     * @param names the name of the column, or more than one if you want.
+     */
+    public Builder includeColumn(String... names) {
+      for (String name : names) {
+        includeColumnNames.add(name);
+        binaryAsStringColumns.add(true);
+      }
+      return this;
+    }
+
+    /**
+     * Include this column.
+     * @param name the name of the column
+     * @param isBinary whether this column is to be read in as binary
+     */
+    public Builder includeColumn(String name, boolean isBinary) {
+      includeColumnNames.add(name);
+      binaryAsStringColumns.add(!isBinary);
+      return this;
+    }
+
+    /**
+     * Include one or more specific columns.  Any column not included will not be read.
+     * @param names the name of the column, or more than one if you want.
+     */
+    public Builder includeColumn(Collection<String> names) {
+      for (String name: names) {
+        includeColumnNames.add(name);
+        binaryAsStringColumns.add(true);
+      }
       return this;
     }
